macOS: remove timeout option
Previously, this software would not build on macOS because timer_t is
not available.

Now, when --timeout or -t option is provided on the command line,
on macOS, an error message is displayed that this option is not
available and the program exits.

See https://github.com/rocasa/ttylog/issues/6